requirements.txt
streamlit
pandas
google-cloud-bigquery
plotly
db-dtypes
